requirements.txt
click==7.0
termcolor==1.1.0
colorama==0.3.7
ruamel.yaml==0.15.6
pandas==0.20.2
bokeh==0.12.6
demandimport==0.3.4
Sphinx==1.6.4
sphinx-click==1.0.3
Jinja2>=2.10.1
python-magic>=0.4
scipy>=0.19.1
Faker>=0.8.16
namedlist>=1.7
GitPython>=2.1.11
pytest==3.6.3
pytest-cov==2.5.1
codecov==2.0.15
numpy==1.14.1
scikit-learn==0.19.0
bkcharts==0.2
statsmodels==0.9.0
importlib-metadata==0.8
path.py==7.7.1
matplotlib==1.5.1
binaryornot==0.4.4
